# misc: make update_licenses.sh faster