@rchincha 2024-08-12
From recent runs, for example https://github.com/project-zot/zot/pull/2594
vast swathes of the license file is getting deleted. So something is broken.