@jkroepke 2024-08-13
I had only such crazy behavior, If I'm running the script before running go mod tidy.